requirements.txt
asgiref==3.8.1 ; python_version >= "3.10" and python_version < "3.14"
async-timeout==5.0.1 ; python_version >= "3.10" and python_full_version < "3.11.3"
certifi==2025.6.15 ; python_version >= "3.10" and python_version < "3.14"
cffi==1.17.1 ; python_version >= "3.10" and python_version < "3.14" and platform_python_implementation != "PyPy"
charset-normalizer==3.4.2 ; python_version >= "3.10" and python_version < "3.14"
click==8.2.1 ; python_version >= "3.10" and python_version < "3.14"
colorama==0.4.6 ; python_version >= "3.10" and python_version < "3.14" and platform_system == "Windows"
cryptography==45.0.4 ; python_version >= "3.10" and python_version < "3.14"
defusedxml==0.7.1 ; python_version >= "3.10" and python_version < "3.14"
dj-database-url==3.0.1 ; python_version >= "3.10" and python_version < "3.14"
django-cors-headers==4.7.0 ; python_version >= "3.10" and python_version < "3.14"
django-filter==25.1 ; python_version >= "3.10" and python_version < "3.14"
django-model-utils==5.0.0 ; python_version >= "3.10" and python_version < "3.14"
django-redis==6.0.0 ; python_version >= "3.10" and python_version < "3.14"
django-rq==3.0.1 ; python_version >= "3.10" and python_version < "3.14"
django-simple-history==3.10.1 ; python_version >= "3.10" and python_version < "3.14"
django==5.2.2 ; python_version >= "3.10" and python_version < "3.14"
djangorestframework-simplejwt==5.5.0 ; python_version >= "3.10" and python_version < "3.14"
djangorestframework==3.16.0 ; python_version >= "3.10" and python_version < "3.14"
djoser==2.3.1 ; python_version >= "3.10" and python_version < "3.14"
idna==3.10 ; python_version >= "3.10" and python_version < "3.14"
injector==0.22.0 ; python_version >= "3.10" and python_version < "3.14"
msal==1.32.3 ; python_version >= "3.10" and python_version < "3.14"
oauthlib==3.3.1 ; python_version >= "3.10" and python_version < "3.14"
office365-rest-python-client==2.6.2 ; python_version >= "3.10" and python_version < "3.14"
psycopg2-binary==2.9.10 ; python_version >= "3.10" and python_version < "3.14"
pycparser==2.22 ; python_version >= "3.10" and python_version < "3.14" and platform_python_implementation != "PyPy"
pyjwt==2.9.0 ; python_version >= "3.10" and python_version < "3.14"
python-decouple==3.8 ; python_version >= "3.10" and python_version < "3.14"
python3-openid==3.2.0 ; python_version >= "3.10" and python_version < "3.14"
pytz==2025.2 ; python_version >= "3.10" and python_version < "3.14"
redis==6.2.0 ; python_version >= "3.10" and python_version < "3.14"
requests-oauthlib==2.0.0 ; python_version >= "3.10" and python_version < "3.14"
requests==2.32.4 ; python_version >= "3.10" and python_version < "3.14"
rq==2.4.0 ; python_version >= "3.10" and python_version < "3.14"
social-auth-app-django==5.4.3 ; python_version >= "3.10" and python_version < "3.14"
social-auth-core==4.6.1 ; python_version >= "3.10" and python_version < "3.14"
sqlparse==0.5.3 ; python_version >= "3.10" and python_version < "3.14"
typing-extensions==4.14.0 ; python_version >= "3.10" and python_version < "3.14"
tzdata==2025.2 ; python_version >= "3.10" and python_version < "3.14" and sys_platform == "win32"
urllib3==2.5.0 ; python_version >= "3.10" and python_version < "3.14"
